added relationships to CartItem model
