+++ app/models/__init__.py
@@ -1,4 +1,4 @@
 from .db import db
 from .user import User
-from .order import Order
+from .cart_item import CartItem
 from .db import environment, SCHEMA
